vit/pyproject.toml
[project]
name = "vit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=3.6.0",
    "flax>=0.10.4",
    "jax==0.4.38",
    "optax>=0.2.5",
    "orbax-checkpoint>=0.11.5",
    "pillow>=11.2.1",
    "polars>=1.31.0",
    "rich>=14.0.0",
]
